backend/manage.py
#!/usr/bin/env python
"""Django's command-line utility for administrative tasks."""
import os
import sys

# Patch PyMySQL before Django loads anything
import pymysql
pymysql.install_as_MySQLdb()

# Fix version info for Django 6.0
import MySQLdb
MySQLdb.VERSION = (2, 2, 1, 'final', 0)
MySQLdb.__version__ = '2.2.1'
MySQLdb.version_info = (2, 2, 1)

def main():
    """Run administrative tasks."""
    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'library_manager.settings')
    try:
        from django.core.management import execute_from_command_line
    except ImportError as exc:
        raise ImportError(
            "Couldn't import Django. Are you sure it's installed and "
            "available on your PYTHONPATH environment variable? Did you "
            "forget to activate a virtual environment?"
        ) from exc
    execute_from_command_line(sys.argv)


if __name__ == '__main__':
    main()
