backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
PyMySQL==1.1.2python-dotenv==1.0.0